Refactor dir for v2
