+++ src/hooks/useEventModal.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { pushAction, popAction, selectIsOpen } from "@/redux/ui/eventModal";
+import { pushAction, popAction, selectIsOpen } from "@/redux/v1/ui/eventModal";
 
 type Push = {
   eventId: string;
