+++ src/hooks/useResource.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { selectUserCalendar } from "@/redux/features/userCalendars";
+import { selectUserCalendar } from "@/redux/v1/features/userCalendars";
 import {
   removeResourceAction,
   pushResourceAction,
-} from "@/redux/features/userCalendars";
+} from "@/redux/v1/features/userCalendars";
 import { initResource } from "@/core/resource/BaseResource";
 
 type IdSet = { calendarId: string; resourceId: string; storyId: string };
