+++ src/hooks/useResourceGroupLabelContentInTemplateCalendar.tsx
@@ -3,8 +3,8 @@ import { useSelector } from "react-redux";
 import {
   selectPrivateCollegeCalendar,
   selectPublicCollegeCalendar,
-} from "@/redux/features/templateCalendar";
-import { selectUserCalendar } from "@/redux/features/userCalendars";
+} from "@/redux/v1/features/templateCalendar";
+import { selectUserCalendar } from "@/redux/v1/features/userCalendars";
 import { updateStory } from "@/core/story/BaseStory";
 import styled from "styled-components";
 import Button from "@mui/material/Button";
