+++ src/hooks/useUser.tsx
@@ -1,6 +1,9 @@
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { updateBirthdayAction, selectUserWithAge } from "@/redux/features/user";
+import {
+  updateBirthdayAction,
+  selectUserWithAge,
+} from "@/redux/v1/features/user";
 
 export const useUser = () => {
   const dispatch = useDispatch();
