+++ src/redux/rootReducer.ts
@@ -1,15 +1,15 @@
 import { combineReducers } from "@reduxjs/toolkit";
 
-import userReducer from "./features/user";
-import templateCalendarReducer from "./features/templateCalendar";
-import templateOptionReducer from "./features/templateOption";
-import userCalendarsReducer from "./features/userCalendars";
-import eventModalReducer from "./ui/eventModal";
-import storyModalReducer from "./ui/storyModal";
-import stepReducer from "./ui/step";
-import previewModalReducer from "./ui/previewModal";
+import userReducer from "@/redux/v1/features/user";
+import templateCalendarReducer from "@/redux/v1/features/templateCalendar";
+import templateOptionReducer from "@/redux/v1/features/templateOption";
+import userCalendarsReducer from "@/redux/v1/features/userCalendars";
+import eventModalReducer from "@/redux/v1/ui/eventModal";
+import storyModalReducer from "@/redux/v1/ui/storyModal";
+import stepReducer from "@/redux/v1/ui/step";
+import previewModalReducer from "@/redux/v1/ui/previewModal";
 
-import { calendarApiV2 } from "./services/calendarApiV2";
+import { calendarApiV2 } from "@/redux/v2/services/calendarApiV2";
 
 const uiReducer = combineReducers({
   storyModal: storyModalReducer,
