+++ src/redux/v1/ui/eventModal.ts
@@ -1,47 +1,49 @@
 import { createSlice, PayloadAction, createSelector } from "@reduxjs/toolkit";
 import { RootState } from "@/redux/rootReducer";
-import { selectStoryByIdFilter } from "@/redux/features/userCalendars";
+import { selectEventByIdFilter } from "@/redux/v1/features/userCalendars";
 
 type State = {
-  story: null | {
+  event: null | {
     calendarId: string;
     storyId: string;
+    eventId: string;
   };
 };
 
 type PushPayload = {
   calendarId: string;
   storyId: string;
+  eventId: string;
 };
 
 type PopPayload = undefined;
 
 const slice = createSlice({
-  name: "storyModal",
+  name: "eventModal",
   initialState: {
-    story: null,
+    event: null,
   } as State,
   reducers: {
     push(state, action: PayloadAction<PushPayload>) {
       const _payload = action.payload;
-      state.story = _payload;
+      state.event = _payload;
     },
     pop(state, _action: PayloadAction<PopPayload>) {
-      state.story = null;
+      state.event = null;
     },
   },
 });
 
-export const selectIsOpen = (state: RootState) => !!state.ui.storyModal.story;
+export const selectIsOpen = (state: RootState) => !!state.ui.eventModal.event;
 
-export const selectStoryModal = (state: RootState) => state.ui.storyModal.story;
+export const selectEventModal = (state: RootState) => state.ui.eventModal.event;
 
-export const selectStory = createSelector(
-  [selectStoryModal, selectStoryByIdFilter],
-  (modal, filter) => {
-    if (!modal) return null;
-    const { calendarId, storyId } = modal;
-    return filter(calendarId, storyId);
+export const selectEvent = createSelector(
+  [selectEventModal, selectEventByIdFilter],
+  (modalInfo, filter) => {
+    if (!modalInfo) return null;
+    const { calendarId, storyId, eventId } = modalInfo;
+    return filter(calendarId, storyId, eventId);
   }
 );
